refactor: appからpage以外のものを出した
