+++ src/app/user/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { Box, Flex, Heading, Button, Container, Avatar, Text } from "@radix-ui/themes";
-import Header from "../components/Header";
+import Header from "../../components/Header";
 import Link from "next/link";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../../contexts/AuthContext";
 import { useRouter } from "next/navigation";
 
 export default function UserPage() {
