+++ src/functions/userService.ts
@@ -1,6 +1,6 @@
 import { db } from "./firebase";
 import { doc, getDoc, setDoc, updateDoc, serverTimestamp } from "firebase/firestore";
-import { User } from "../../types";
+import { User } from "../types";
 
 /**
  * ユーザーの作成
